refactor(business_profile): copy `AuthenticationConnectorDetails` to `diesel_models`
